Bugfix for call of xfoil_scale_LE_radius
There seems to be a problem, if "new_foil" is an input-parameter and an utput-parameter at the same time.
When entering the function "xfoil_scale_LE_radius" the content of "new_foil" gets invalid.
If using "new_foil_1" as an input-parameter and "new_foil_2" as an output parameter there is no seg-fault anymore.