Changes in convergence_profile "fastest"->"quick_camb_thick"
Adaption of modes in case of convergence-profile "quick_camb_thick" only to minimize retroactive effects on strak-machine.
Convergence-profile "quick_camb_thick" finds more solutions especially for the first strak airfoil and this has an impact on the creation of the strak.
The input-file has to be adapted for "quick_camb_thick"