Re-Implementation of automatic adaption of initialPerturb
The value of initial perturb will be adapted, according to the "challange" for the optimizer. This means, if the change in Re-number between the root-airfoil and the strak-airfoils is big, also a bigger initial-perturb will be needed. For smaller changes in Re-number, the initial perturb will be set to smaller numbers.